Remove PromptSkimming from the system
The ProdAgent is retired, this workflow is not valid anymore.
The OpenRunningRequests provide support for PromptSkim without an
actual relation to the T0 and iyt is also included in the PromptReco
spec.

This is part of the effor to improve the argument validation.
