--- src/python/WMComponent/PromptSkimScheduler/__init__.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-"""
-__init__
-
-"""
-__all__ = []
